requirements.txt
# 增强版文档提取器依赖包
# 核心依赖
playwright>=1.40.0
Pillow>=10.0.0
colorama>=0.4.6

# 可选依赖
requests>=2.31.0          # HTTP请求
beautifulsoup4>=4.12.0   # HTML解析
numpy>=1.24.0            # 数值计算
pandas>=2.0.0            # 数据处理

# 开发依赖 (可选)
pytest>=7.4.0            # 测试框架
black>=23.0.0            # 代码格式化
flake8>=6.0.0            # 代码检查
mypy>=1.5.0              # 类型检查